+++ R/AvoPhylo.R
@@ -1,6 +1,8 @@
 utils::globalVariables(c("phylo_id", "group",
                          "BLFamilyLatin", "Order",
-                         "Clade", "i", "stopCluster"))
+                         "Clade", "i", "stopCluster",
+                         "type", "jetz_family",
+                         "jetz_order"))
 
 #' AvoPhylo: Building phylogenies based on the AvoTrex extinct bird trait
 #' database and BirdTree backbone
@@ -302,7 +304,7 @@ AvoPhylo <- function(
       for(p in 1:length(groups)){
         
         shuff2 <- dplyr::filter(shuff, group == groups[p])
-        shuff2 <-  shuff2[sample(1:nrow(shuff2)), ]
+        shuff2 <- shuff2[sample(1:nrow(shuff2)), ]
         ex <- rbind(ex, shuff2)
         
       }
@@ -310,8 +312,20 @@ AvoPhylo <- function(
       row.names(ex) <- 1:nrow(ex)
       }#eo if xs
       
+      #create copy which is used in the check_ext() function;
+      #this adds extinct order, family and genus names to the
+      #jetz columns
+      ex2 <- ex
+      Bl_levs <- c("order", "family", "genus")
+      BT_levs <- c("jetz_order", "jetz_family", "jetz_genus")
+      for (z in 1:length(Bl_levs)){
+        wlevs <- which(ex2[,BT_levs[z]] == "Extinct")
+        ex2[wlevs,BT_levs[z]] <- ex2[wlevs,Bl_levs[z]]
+      }
+      ex2$jetz_order <- toupper(ex2$jetz_order)
+      
       # For each extinct species find the optimum place to bind 
-      for(j in 1:nrow(ex)){
+      for (j in 1:nrow(ex)){
         
         #First check if species is time_fixed, and if so, extract the time point
         #info. If not extract the species' per code and set the PER_VAL and
@@ -351,9 +365,9 @@ AvoPhylo <- function(
         # write.csv(vec, paste0("Broke/broke_", i, ".csv"), row.names = F)
         
         # AP = Already present - Nothing needs to be done  
-        if(ex$type[j] == "AP"){
+        if (ex$type[j] == "AP"){
           next
-        } else if(ex$type[j] == "S"){
+        } else if (ex$type[j] == "S"){
           ## Scenario 1.1: Add target species as a single sister of species X 
           ## - S (SISTER SPECIES)
           
@@ -361,37 +375,31 @@ AvoPhylo <- function(
           nodeX <- which(ctree$tip.label == paste0(ex$sister_genus[j], 
                                                    "_", ex$sister_species[j])) 
           
-        } else if(ex$type[j] %in% c("SSG", "SGG", "SGG2", "SFG", "SOG")){
+        } else if (ex$type[j] %in% c("SSG", "SGG", "SGG2", "SFG", "SOG")){
           ## Scenario 1.2: Add species as a sister (outgroup) of a group of
           ## species ## - SSG (SISTER SPECIES GROUP) & SGG (SISTER GENUS GROUP)
           ## & SFG (SISTER FAMILY GROUP) & SOG (SISTER ORDER GROUP)
           
-          if(ex$type[j] == "SSG"){
+          if (ex$type[j] == "SSG"){
             
             # Separate the species in the "sister_species_group" column
             sp <- stringr::str_split(ex$sister_species_group[j], pattern = ";")
-            spv <- vector()
-            
-            for(x in 1:length(sp[[1]])){
-              
-              spv <- c(spv, paste0(sp[[1]][x]))
-              
-            } # Make the species group as a vector
-            
+            spv <- sp[[1]]
+
             # This selects the most recent common ancestor for the group of species
             nodeX <- getMRCA(ctree, spv) 
-          }else if(ex$type[j] == "SGG"){
+          } else if (ex$type[j] == "SGG"){
             
             # If only one species is present within the genus, then make a
             # sister to that species
-            if(length(ctree$tip.label[grep(paste0(ex$sister_genus[j], "_"), 
+            if (length(ctree$tip.label[grep(paste0(ex$sister_genus[j], "_"), 
                                             ctree$tip.label)]) == 1){
               # Get the tip location for the sister sp.
               nodeX <- which(ctree$tip.label == 
                                ctree$tip.label[grep(paste0(ex$sister_genus[j], 
                                                            "_"), ctree$tip.label)]) 
 
-            }else{
+            } else {
 
               # Get most recent common ancestor of genus 
               nodeX <- getMRCA(ctree, 
@@ -399,68 +407,57 @@ AvoPhylo <- function(
                                                            "_"), ctree$tip.label)]) 
  
             }
-          }else if(ex$type[j] == "SGG2"){
+          } else if (ex$type[j] == "SGG2"){
             
             sp <- stringr::str_split(ex$sister_genus[j], pattern = ";")
             spv <- vector()
             
             # Get all the species in the genera
-            for(x in 1:length(sp[[1]])){
+            for (x in 1:length(sp[[1]])){
               
               spv <- c(spv, ctree$tip.label[grep(paste0(sp[[1]][x], "_"),
                                                  ctree$tip.label)])
               
             }
-            
             # Get most recent common ancestor of species group
             nodeX <- getMRCA(ctree, spv)  
-          }else if(ex$type[j] == "SFG"){
+          } else if (ex$type[j] == "SFG"){
             
             # Get all species within the family 
-            fam <- dplyr::filter(tax, BLFamilyLatin == ex$sister_family[j])
-            spv <- vector()
-            
-            for(x in 1:nrow(fam)){
-              
-              spv <- c(spv, paste0(fam$Genus[x], "_", fam$Species[x]))
-              
-            } # Make the species group as a vector
+            fam <- dplyr::filter(tax, 
+                                 BLFamilyLatin == ex$sister_family[j])
+            spv <- fam$TipLabel
+            #add in extinct sp already grafted
+            spv <- check_ext(ex2, level = ex$sister_family[j], 
+                             ctree, spv, code = "SFG")
             
             # This selects the most recent common ancestor for the group of species
             nodeX <- getMRCA(ctree, spv)
-          }else if(ex$type[j] == "SOG"){
+          } else if (ex$type[j] == "SOG"){
             
             # Get all species within the family 
             ord <- dplyr::filter(tax, Order == ex$sister_order[j])
-            spv <- vector()
-            
-            for(x in 1:nrow(ord)){
-              
-              spv <- c(spv, paste0(ord$Genus[x], "_", ord$Species[x]))
-              
-            } # Make the species group as a vector
-            
+            spv <- ord$TipLabel
+            #add in extinct sp already grafted
+            spv <- check_ext(ex2, level = ex$sister_order[j], 
+                             ctree, spv, code = "SOG")
+         
             # This selects the most recent common ancestor for the group of species
             nodeX <- getMRCA(ctree, spv) 
           }
           
-        } else if(ex$type[j] %in% c("RSG", "RGG", "RGG2", "RCG", "RFG", "ROG")){
+        } else if (ex$type[j] %in% c("RSG", "RGG", "RGG2", "RCG", "RFG", "ROG")){
           ## Scenario 2.1: Add species randomly within of a group of species ##
           ## - RSG (RANDOM SPECIES GROUP) & RGG (RANDOM GENUS GROUP) & RFG
           ## (RANDOM FAMILY GROUP) & ROG (RANDOM ORDER GROUP)
           
-          if(ex$type[j] == "RSG"){
+          if (ex$type[j] == "RSG"){
             
             # Separate the species in the "sister_species_group" column
-            sp <- stringr::str_split(ex$sister_species_group[j], pattern = ";")
-            spv <- vector()
-            
-            for(x in 1:length(sp[[1]])){
-              
-              spv <- c(spv, paste0(sp[[1]][x]))
-              
-            } 
-            
+            sp <- stringr::str_split(ex$sister_species_group[j], 
+                                     pattern = ";")
+            spv <- sp[[1]]
+    
             ## Randomly select one of the species from the listed group. If it
             ## is not present in the tree yet, remove from the list and select
             ## again. Will break if all species have been attempted and there
@@ -471,45 +468,39 @@ AvoPhylo <- function(
               # Get the tip location for the sister sp.
               nodeX <- which(ctree$tip.label == spv2) 
               # Break if there is a node value
-              if(length(nodeX) != 0) break else{
+              if (length(nodeX) != 0) break else{
                 spv <- spv[!spv == spv2]
               }
-              if(length(spv) == 0) break
+              if (length(spv) == 0) break
               
             }
             
             ## Check if the node still is length zero
-            if(length(nodeX) == 0){message(paste0("Node is still zero length for ", 
+            if (length(nodeX) == 0){message(paste0("Node is still zero length for ", 
                                                 ex$type[j], " for species ", 
                                                 ex$species[j],
                                                 " (row ", j, ") after random species selection."))}
 
-          }else if(ex$type[j] == "RGG"){
+          } else if (ex$type[j] == "RGG"){
             
             # Get all the species in the genus
-            sp <- stringr::str_split(ex$sister_genus[j], pattern = ";")
-            spv <- vector()
-            
-            # Get all the species in the genera
-            for(x in 1:length(sp[[1]])){
-              
-              spv <- c(spv, ctree$tip.label[grep(paste0(sp[[1]][x], "_"), 
-                                                 ctree$tip.label)])
-              
-            }
+            sp <- ex$sister_genus[j]
+            spv <- ctree$tip.label[grep(paste0(sp, "_"), 
+                                        ctree$tip.label)]
             
             #Randomly select one of the species
             spv2 <- sample(spv, 1)
             
             # Get the tip location for the sister sp.
             nodeX <- which(ctree$tip.label == spv2) 
-          }else if(ex$type[j] == "RGG2"){
+          } else if (ex$type[j] == "RGG2"){
             
             sp <- stringr::str_split(ex$sister_genus[j], pattern = ";")
             spv <- vector()
             
-            # Get all the species in the genera
-            for(x in 1:length(sp[[1]])){
+            # Get all the species in the genera (incl extinct 
+            # already grafted)
+            for (x in 1:length(sp[[1]])){
               
               spv <- c(spv, ctree$tip.label[grep(paste0(sp[[1]][x], "_"), 
                                                  ctree$tip.label)])
@@ -521,65 +512,63 @@ AvoPhylo <- function(
             
             # Get the tip location for the sister sp.
             nodeX <- which(ctree$tip.label == spv2) 
-          }else if(ex$type[j] == "RFG"){
+          } else if (ex$type[j] == "RFG"){
             
             # Get all species within the family 
-            fam <- dplyr::filter(tax, BLFamilyLatin == ex$sister_family[j])
-            spv <- vector()
-            
-            for(x in 1:nrow(fam)){
-              
-              spv <- c(spv, paste0(fam$Genus[x], "_", fam$Species[x]))
-              
-            } # Make the species group as a vector
-            
+            fam <- dplyr::filter(tax, 
+                                 BLFamilyLatin == ex$sister_family[j])
+            spv <- fam$TipLabel
+            #add in extinct sp already grafted in this family
+            spv <- check_ext(ex2, level = ex$sister_family[j], 
+                             ctree, spv, code = "RFG")
+
             #Randomly select one of the species
             spv2 <- sample(spv, 1)
             
             # Get the tip location for the sister sp.
             nodeX <- which(ctree$tip.label == spv2) 
-          }else if(ex$type[j] == "RCG"){
-            
-            # Get all species within the family 
-            fam <- dplyr::filter(tax, Clade == ex$sister_clade[j])
-            spv <- vector()
-            
-            for(x in 1:nrow(fam)){
-              
-              spv <- c(spv, paste0(fam$Genus[x], "_", fam$Species[x]))
-              
-            } # Make the species group as a vector
-            
-            #Randomly select one of the species
-            spv2 <- sample(spv, 1)
-            
-            # Get the tip location for the sister sp.
-            nodeX <- which(ctree$tip.label == spv2) 
-          }else if(ex$type[j] == "ROG"){
-            
-            # Get all species within the family 
-            ord <- dplyr::filter(tax, Order == ex$sister_family[j])
-            spv <- vector()
-            
-            for(x in 1:nrow(fam)){
-              
-              spv <- c(spv, paste0(ord$Genus[x], "_", ord$Species[x]))
-              
-            } # Make the species group as a vector
-            
-            #Randomly select one of the species
-            spv2 <- sample(spv, 1)
-            
-            # Get the tip location for the sister sp.
-            nodeX <- which(ctree$tip.label == spv2) 
-          }
-          
-        } else if(ex$type[j] %in% c("RSGG", "RSGG2")){
+           } # else if (ex$type[j] == "RCG"){
+          # 
+          #   # Get all species within the family
+          #   fam <- dplyr::filter(tax, Clade == ex$sister_clade[j])
+          #   spv <- vector()
+          # 
+          #   for (x in 1:nrow(fam)){
+          # 
+          #     spv <- c(spv, paste0(fam$Genus[x], "_", fam$Species[x]))
+          # 
+          #   } # Make the species group as a vector
+          # 
+          #   #Randomly select one of the species
+          #   spv2 <- sample(spv, 1)
+          # 
+          #   # Get the tip location for the sister sp.
+          #   nodeX <- which(ctree$tip.label == spv2)
+          # } else if (ex$type[j] == "ROG"){
+          # 
+          #   # Get all species within the family
+          #   ord <- dplyr::filter(tax, Order == ex$sister_family[j])
+          #   spv <- vector()
+          # 
+          #   for(x in 1:nrow(fam)){
+          # 
+          #     spv <- c(spv, paste0(ord$Genus[x], "_", ord$Species[x]))
+          # 
+          #   } # Make the species group as a vector
+          # 
+          #   #Randomly select one of the species
+          #   spv2 <- sample(spv, 1)
+          # 
+          #   # Get the tip location for the sister sp.
+          #   nodeX <- which(ctree$tip.label == spv2)
+          # }
+          # 
+        } else if (ex$type[j] %in% c("RSGG", "RSGG2")){
           ## Scenario 3.1: Add species as a sister to a genus selected randomly
           ## from a supplied group of genera (RSGG) or a random genus from a
           ## supplied family (RSGG2)
           
-          if(ex$type[j] == "RSGG"){
+          if (ex$type[j] == "RSGG"){
             
             #Split the supplied genera
             sp <- stringr::str_split(ex$sister_genus[j], pattern = ";")
@@ -587,7 +576,7 @@ AvoPhylo <- function(
             #Randomly select one of the genera
             spv2 <- sample(sp[[1]], 1)
             
-            if(length(ctree$tip.label[grep(paste0(spv2, "_"), 
+            if (length(ctree$tip.label[grep(paste0(spv2, "_"), 
                                            ctree$tip.label)]) == 1){
               
               # Get the tip location for single species in the genus
@@ -595,37 +584,29 @@ AvoPhylo <- function(
                                ctree$tip.label[grep(paste0(spv2, "_"), 
                                                     ctree$tip.label)]) 
               
-            }else{
+            } else {
               # Get most recent common ancestor of genus 
               nodeX <- getMRCA(ctree, ctree$tip.label[grep(paste0(spv2, "_"), 
                                                            ctree$tip.label)]) 
             }
           }
           
-          if(ex$type[j] == "RSGG2"){
+          if (ex$type[j] == "RSGG2"){
             
             # Get all genera within the family 
             fam <- dplyr::filter(tax, BLFamilyLatin == ex$sister_family[j])
-            spv <- vector()
-            
-            for(x in 1:nrow(fam)){
-              
-              spv <- c(spv, paste0(fam$Genus[x]))
-              
-            } # Make the species group as a vector
-            
-            spv <- unique(spv)
-            
+            spv <- unique(fam$Genus)
+
             #Randomly select one of the genera
             spv2 <- sample(spv, 1)
             
-            if(length(ctree$tip.label[grep(paste0(spv2, "_"), 
+            if (length(ctree$tip.label[grep(paste0(spv2, "_"), 
                                            ctree$tip.label)]) == 1){
               # Get the tip location for single species in the genus
               nodeX <- which(ctree$tip.label == 
                                ctree$tip.label[grep(paste0(spv2, "_"), 
                                                     ctree$tip.label)]) 
-            } else{
+            } else {
               # Get most recent common ancestor of genus 
               nodeX <- getMRCA(ctree,ctree$tip.label[grep(paste0(spv2, "_"), 
                                                           ctree$tip.label)]) 
@@ -674,3 +655,26 @@ AvoPhylo <- function(
   return(ctreesComplete)
   
 }
+
+#Internal function to check which extinct species from
+#a given family or order have already been grafted, and
+#to select those that have
+#' @importFrom dplyr filter
+check_ext <- function(ex2, level, ctree, spv, code){
+  if (code == "RFG" | code == "SFG"){
+    spvE <- dplyr::filter(ex2, 
+                          type != "AP",
+                          jetz_family == level)
+  } else if (code == "SOG"){
+    spvE <- dplyr::filter(ex2, 
+                          type != "AP",
+                          jetz_order == level)
+  }
+  if (nrow(spvE) > 0){
+    if (any(spvE$species %in% ctree$tip.label)){
+      wspvE <- which(spvE$species %in% ctree$tip.label)
+      spv <- c(spv, spvE$species[wspvE])
+    }
+  }#eo if 0
+  return(spv)
+}#eo function