+++ NEWS.md
@@ -1,4 +1,12 @@
 
+
+## Version 1.3.0 
+  *Updating database to 610 extinct species
+  *Changing SOG, SFG and RFG grafting to also include
+  extinct sp. in the order / family that have already been
+  grafted (using check_ext() function)
+  *small bug correction in AgeBind
+
 ## Version 1.2.0 
   *small bug corrections in the plot and AgeBind functions
 
@@ -9,7 +17,6 @@
   * AgeBind() internal function added which allows for species
     to be grafted at a specific time point along a branch.
   * AvoBind() and AvoPhylo() updated to work with AgeBind()
-  
   
 ## Version 1.0.0